requirements.txt
gensim>=4.3.0
tiktoken>=0.5.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
plotly>=5.18.0
numpy>=1.24.0
umap-learn>=0.5.0
